Add implicit ssh workaround
OpenNebula has hardwired SSH port 22 - frontend and nodes has to be able
to connect to each other via this port.

The opennebula-node container running with this docker-compose.yml is
already fixed by simply running its ssh on different port. But for all
other nodes outside of our docker-compose we need to fix this problem.

This commit workarounds this issue thnaks to the ProxyCommand and no
further hacks on the nodes should be needed *BUT*...

IMPORTANT:

For this tweak to work the host (where the opennebula-frontend
container is running) must have created a user oneadmin with proper
authorized_keys file containing our oneadmin's public ssh key!

BEWARE:

The trick is quite nasty and much more sensible would be to just publish
the inner OpenNebula's SSH port as the standard 22 and change the host's
SSH port instead - the workaround count with that too - so if you can do
it like that, thanks.

NOTE:

Although deploy works - undeploy will still fail if SSH port is not set
to the standard 22 - workaround is not yet complete...

Signed-off-by: Petr Ospalý <pospaly@opennebula.io>